# mobile/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "GH Trust MFB API",
  "description": "GH Trust International Ltd \u2014 Microfinance Banking API",
  "version": "0.2.0"
 },
 "paths": {
...
  "/api/v1/wallet": {
   "get": {
    "tags": [
     "Wallet"
    ],
    "summary": "Get Wallet",
    "operationId": "get_wallet_api_v1_wallet_get",
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/WalletSummaryResponse"
        }
       }
      }
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ]
   }
  },
  "/api/v1/wallet/fund": {
   "post": {
    "tags": [
     "Wallet"
    ],
    "summary": "Create Wallet Funding Session",
    "description": "Zest: generate a temporary virtual account for bank transfer (expires ~5 minutes).",
    "operationId": "create_wallet_funding_session_api_v1_wallet_fund_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/WalletFundRequest"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/WalletFundSessionResponse"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ]
   }
  },
  "/api/v1/wallet/payout-account": {
   "post": {
    "tags": [
     "Wallet"
    ],
    "summary": "Update Payout Account",
    "operationId": "update_payout_account_api_v1_wallet_payout_account_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/UpdatePayoutAccountRequest"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ]
   }
  },
  "/api/v1/wallet/withdraw": {
   "post": {
    "tags": [
     "Wallet"
    ],
    "summary": "Request Withdrawal",
    "description": "Requires an `Idempotency-Key` header: a retried request with the same key returns the original withdrawal instead of creating a second one.",
    "operationId": "request_withdrawal_api_v1_wallet_withdraw_post",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "Idempotency-Key",
      "in": "header",
      "required": true,
      "schema": {
       "type": "string",
       "minLength": 8,
       "maxLength": 64,
       "title": "Idempotency-Key"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/WithdrawRequest"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/WithdrawalResponse"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "UpdatePayoutAccountRequest": {
    "properties": {
     "bank_code": {
      "type": "string",
      "maxLength": 10,
      "minLength": 3,
      "title": "Bank Code"
     },
     "account_number": {
      "type": "string",
      "maxLength": 10,
      "minLength": 10,
      "title": "Account Number"
     },
     "account_name": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 200
       },
       {
        "type": "null"
       }
      ],
      "title": "Account Name"
     },
     "bank_name": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 100
       },
       {
        "type": "null"
       }
      ],
      "title": "Bank Name"
     },
     "transaction_pin": {
      "type": "string",
      "maxLength": 4,
      "minLength": 4,
      "pattern": "^\\d+$",
      "title": "Transaction Pin",
      "description": "Customer's 4-digit transaction PIN"
     }
    },
    "type": "object",
    "required": [
     "bank_code",
     "account_number",
     "transaction_pin"
    ],
    "title": "UpdatePayoutAccountRequest"
   },
...
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   },
...
   "WalletFundRequest": {
    "properties": {
     "amount": {
      "anyOf": [
       {
        "type": "number",
        "exclusiveMinimum": 0.0
       },
       {
        "type": "string"
       }
      ],
      "title": "Amount"
     }
    },
    "type": "object",
    "required": [
     "amount"
    ],
    "title": "WalletFundRequest"
   },
   "WalletFundSessionResponse": {
    "properties": {
     "transaction_ref": {
      "type": "string",
      "title": "Transaction Ref"
     },
     "account_number": {
      "type": "string",
      "title": "Account Number"
     },
     "account_name": {
      "type": "string",
      "title": "Account Name"
     },
     "bank_name": {
      "type": "string",
      "title": "Bank Name"
     },
     "amount": {
      "type": "number",
      "title": "Amount"
     },
     "currency": {
      "type": "string",
      "title": "Currency"
     },
     "expires_in_minutes": {
      "type": "integer",
      "title": "Expires In Minutes"
     },
     "message": {
      "type": "string",
      "title": "Message",
      "default": "Transfer to this account within the expiry window"
     }
    },
    "type": "object",
    "required": [
     "transaction_ref",
     "account_number",
     "account_name",
     "bank_name",
     "amount",
     "currency",
     "expires_in_minutes"
    ],
    "title": "WalletFundSessionResponse"
   },
   "WalletSummaryResponse": {
    "properties": {
     "available_balance": {
      "type": "number",
      "title": "Available Balance"
     },
     "locked_balance": {
      "type": "number",
      "title": "Locked Balance"
     },
     "currency": {
      "type": "string",
      "title": "Currency"
     },
     "dva_status": {
      "type": "string",
      "title": "Dva Status"
     },
     "dva_account_number": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Dva Account Number"
     },
     "dva_bank_name": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Dva Bank Name"
     },
     "paystack_customer_code": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Paystack Customer Code"
     },
     "payment_provider": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Payment Provider"
     },
     "funding_mode": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Funding Mode"
     }
    },
    "type": "object",
    "required": [
     "available_balance",
     "locked_balance",
     "currency",
     "dva_status"
    ],
    "title": "WalletSummaryResponse"
   },
...
   "WithdrawRequest": {
    "properties": {
     "amount": {
      "anyOf": [
       {
        "type": "number",
        "exclusiveMinimum": 0.0
       },
       {
        "type": "string"
       }
      ],
      "title": "Amount"
     },
     "transaction_pin": {
      "type": "string",
      "maxLength": 4,
      "minLength": 4,
      "pattern": "^\\d+$",
      "title": "Transaction Pin",
      "description": "Customer's 4-digit transaction PIN"
     }
    },
    "type": "object",
    "required": [
     "amount",
     "transaction_pin"
    ],
    "title": "WithdrawRequest"
   },
   "WithdrawalResponse": {
    "properties": {
     "id": {
      "type": "string",
      "title": "Id"
     },
     "amount": {
      "type": "number",
      "title": "Amount"
     },
     "status": {
      "type": "string",
      "title": "Status"
     },
     "transfer_reference": {
      "type": "string",
      "title": "Transfer Reference"
     },
     "message": {
      "type": "string",
      "title": "Message",
      "default": "Withdrawal queued for processing"
     }
    },
    "type": "object",
    "required": [
     "id",
     "amount",
     "status",
     "transfer_reference"
    ],
    "title": "WithdrawalResponse"
   },
...
  },
  "securitySchemes": {
   "HTTPBearer": {
    "type": "http",
    "scheme": "bearer"
   }
  }
 }
}